Cleaned up the code, added colors
